🚧 Removal of CardData child references
This is the start of the branch and a breaking change and current code cannot be compiled at this commit. Code is being revised to fit the new story-tree paradigm.
Code is only commented out, so it can be referenced while rewriting the code (It should be removed completely later).